+++ app/components/hoodie.admin/hoodie.admin.js
@@ -98,7 +98,7 @@ Hoodie.AdminApp = (function() {
     var defer, info;
     defer = this.hoodie.defer();
     info = {
-      name: "appName here"
+      name: "appName (not implemented yet)"
     };
     window.setTimeout(function() {
       return defer.resolve(info);
@@ -192,9 +192,6 @@ Hoodie.AdminModules = (function(_super) {
     if (!moduleName) {
       moduleName = type;
     }
-    if (moduleName === 'module') {
-      debugger;
-    }
     return AdminModules.__super__.find.call(this, "module", moduleName);
   };
 
@@ -203,7 +200,7 @@ Hoodie.AdminModules = (function(_super) {
   };
 
   AdminModules.prototype.update = function(moduleName, config) {
-    debugger;    return AdminModules.__super__.update.call(this, "module", moduleName, config);
+    return AdminModules.__super__.update.call(this, "module", moduleName, config);
   };
 
   AdminModules.prototype.getConfig = function(moduleName) {
@@ -429,9 +426,11 @@ Hoodie.Admin = (function() {
       options = {};
     }
     $.extend(options, {
-      name: storeName,
-      baseUrl: this.baseUrl
+      name: storeName
     });
+    if (this.baseUrl !== this.hoodie.baseUrl) {
+      options.baseUrl = this.baseUrl;
+    }
     return new Hoodie.Remote(this.hoodie, options);
   };